Update config with separate entry points for beamer/website
