+++ intranet-client/config/webpack.config.js
@@ -23,6 +23,7 @@ const getClientEnvironment = require('./env');
 const ModuleNotFoundPlugin = require('react-dev-utils/ModuleNotFoundPlugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin-alt');
 const typescriptFormatter = require('react-dev-utils/typescriptFormatter');
+const WebpackAssetsManifest = require('webpack-assets-manifest');
 
 
 // Source maps are resource heavy and can cause out of memory issue for large source files.
@@ -124,25 +125,14 @@ module.exports = function(webpackEnv) {
       : isEnvDevelopment && 'cheap-module-source-map',
     // These are the "entry points" to our application.
     // This means they will be the "root" imports that are included in JS bundle.
-    entry: [
-      // Include an alternative client for WebpackDevServer. A client's job is to
-      // connect to WebpackDevServer by a socket and get notified about changes.
-      // When you save a file, the client will either apply hot updates (in case
-      // of CSS changes), or refresh the page (in case of JS changes). When you
-      // make a syntax error, this client will display a syntax error overlay.
-      // Note: instead of the default WebpackDevServer client, we use a custom one
-      // to bring better experience for Create React App users. You can replace
-      // the line below with these two lines if you prefer the stock client:
-      // require.resolve('webpack-dev-server/client') + '?/',
-      // require.resolve('webpack/hot/dev-server'),
-      isEnvDevelopment &&
-        require.resolve('react-dev-utils/webpackHotDevClient'),
-      // Finally, this is your app's code:
-      paths.appIndexJs,
-      // We include the app code last so that if there is a runtime error during
-      // initialization, it doesn't blow up the WebpackDevServer client, and
-      // changing JS code would still trigger a refresh.
-    ].filter(Boolean),
+    entry: {
+      beamer: [
+        path.resolve(paths.appSrc, 'beamer/index'),
+      ],
+      website: [
+        path.resolve(paths.appSrc, 'website/index'),
+      ],
+    },
     output: {
       // The build folder.
       path: isEnvProduction ? paths.appBuild : undefined,
@@ -577,7 +567,7 @@ module.exports = function(webpackEnv) {
             module: 'esnext',
             moduleResolution: 'node',
             resolveJsonModule: true,
-            isolatedModules: true,
+            isolatedModules: false,
             noEmit: true,
             jsx: 'preserve',
           },
@@ -592,6 +582,10 @@ module.exports = function(webpackEnv) {
           watch: paths.appSrc,
           silent: true,
           formatter: typescriptFormatter,
+        }),
+
+        new WebpackAssetsManifest({
+          entrypoints: true,
         }),
     ].filter(Boolean),
     // Some libraries import Node modules but don't use them in the browser.
