+++ intranet-client/package.json
@@ -79,7 +79,8 @@
     "@types/socket.io-client": "^1.4.32",
     "node-sass": "^4.11.0",
     "prettier": "^1.16.3",
-    "typescript": "^3.3.1"
+    "typescript": "^3.3.1",
+    "webpack-assets-manifest": "^3.1.1"
   },
   "jest": {
     "collectCoverageFrom": [
