+++ intranet-client/src/beamer/_variables.scss
