+++ intranet-client/src/beamer/index.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import 'variables';
 
 @import url('https://fonts.googleapis.com/css?family=Work+Sans:400,600');
