+++ intranet-client/src/beamer/panels/Panel.scss
