+++ intranet-client/src/beamer/panels/Panel.tsx
