+++ intranet-client/src/beamer/panels/PanelGroup.scss
