+++ intranet-client/src/beamer/panels/PanelGroup.tsx
