+++ intranet-client/src/beamer/panels/activity/ActivityPanel.scss
