+++ intranet-client/src/beamer/panels/shoutbox/Shoutbox.tsx
