+++ intranet-client/src/beamer/panels/shoutbox/ShoutboxMessage.scss
