+++ intranet-client/src/beamer/panels/slides/SlidesPanel.scss
