+++ intranet-client/src/beamer/panels/slides/data/gepwnage-logo.png
