+++ intranet-client/src/beamer/panels/slides/data/letstalk-logo.png
