+++ intranet-client/src/beamer/panels/time/TimePanel.scss
