+++ intranet-client/src/beamer/panels/time/useTime.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { useInterval } from '../../helpers/useInterval';
+import { useInterval } from '../../../helpers/useInterval';
 
 export default function useTime(delay = 1000): [Date] {
     const [time, setTime] = React.useState(new Date());
