+++ intranet-client/src/website/index.ts
@@ -1,0 +1,3 @@
+import './index.scss';
+
+console.log('Hello, World!');
